# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chatbot",
        "description": "API para la gestión de usuarios, chats y mensajes de un chatbot. Documentación en español.",
        "version": "1"
    },
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión de usuario",
                "description": "Autentica a un usuario y genera un token de acceso.",
                "operationId": "dd51a484b7fa0f97c7680fdbde14cf5f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secreto123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserDTO"
                                        },
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    },
                    "403": {
                        "description": "Cuenta no activa"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener información del usuario autenticado",
                "description": "Devuelve los datos del usuario autenticado y sus tokens.",
                "operationId": "b5de204f5e452ef51327c691aa99e77c",
                "responses": {
                    "200": {
                        "description": "Usuario autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/UserDTO"
                                        },
                                        "tokens": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Revoca el token actual y cierra la sesión del usuario autenticado.",
                "operationId": "df9bdc6c7a6d9f20b983526b2ec0f698",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Refrescar token de acceso",
                "description": "Genera un nuevo token de acceso y revoca el anterior.",
                "operationId": "30c20c52f5f72828d8c531fd7c9ea3e7",
                "responses": {
                    "200": {
                        "description": "Token refrescado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "type": "string"
                                        },
                                        "expires_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/tokens": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Listar tokens del usuario",
                "description": "Devuelve la lista de tokens del usuario autenticado.",
                "operationId": "13613cb0146f64ecd197e95e0ecf5df8",
                "responses": {
                    "200": {
                        "description": "Lista de tokens",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tokens": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/tokens/{tokenId}": {
            "delete": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Revocar token específico",
                "description": "Revoca un token específico del usuario autenticado.",
                "operationId": "abd12d9be1adb3a75776de0a7b5bd633",
                "parameters": [
                    {
                        "name": "tokenId",
                        "in": "path",
                        "description": "ID del token a revocar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token revocado correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserDTO": {
                "title": "Usuario",
                "description": "DTO que representa a un usuario del sistema.",
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del usuario",
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "description": "Correo electrónico del usuario",
                        "type": "string",
                        "format": "email",
                        "example": "juan@example.com"
                    },
                    "password": {
                        "description": "Contraseña del usuario (solo para registro o actualización)",
                        "type": "string",
                        "format": "password",
                        "example": "secreto123"
                    },
                    "role": {
                        "description": "Rol asignado al usuario",
                        "type": "string",
                        "example": "admin"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación mediante token Bearer. Ingrese 'Bearer {token}' en el encabezado Authorization.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}